# api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SYA Scheduler Function API",
    "version": "0.1.0",
    "description": "Scheduler-compatible task decomposition and rolling-wave planning API for SYA."
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "description": "Local SYA function runtime. Electron assigns the port dynamically.",
      "variables": {
        "port": {
          "default": "0"
        }
      }
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Scheduler function health check",
        "responses": {
          "200": {
            "description": "Runtime is healthy",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/manifest": {
      "get": {
        "operationId": "manifest",
        "summary": "Scheduler function manifest",
        "responses": {
          "200": {
            "description": "Function manifest",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/decompose": {
      "post": {
        "operationId": "decompose",
        "summary": "Decompose natural-language input into tasks and timeline events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DecomposeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decomposition result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/tasks": {
      "get": {
        "operationId": "tasks.list",
        "summary": "List tasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "in_progress", "completed", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/tasks/{taskId}": {
      "get": {
        "operationId": "tasks.get",
        "summary": "Get one task",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "tasks.update",
        "summary": "Update a task",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "tasks.delete",
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/timeline": {
      "get": {
        "operationId": "timeline.get",
        "summary": "Get timeline view",
        "responses": {
          "200": {
            "description": "Timeline data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/tasks/reorder": {
      "post": {
        "operationId": "tasks.reorder",
        "summary": "Reorder tasks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReorderTasksRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reorder result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/stats": {
      "get": {
        "operationId": "stats.get",
        "summary": "Get daily task statistics",
        "responses": {
          "200": {
            "description": "Daily stats",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/config": {
      "get": {
        "operationId": "config.get",
        "summary": "Get scheduler configuration",
        "responses": {
          "200": {
            "description": "Scheduler configuration",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "config.update",
        "summary": "Update scheduler configuration",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated scheduler configuration",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/events": {
      "get": {
        "operationId": "events",
        "summary": "Server-sent event stream",
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/actions": {
      "get": {
        "operationId": "actions.list",
        "summary": "List callable actions",
        "responses": {
          "200": {
            "description": "Action list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FunctionResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FunctionResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean" },
          "data": {},
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": {}
            }
          }
        }
      },
      "DecomposeRequest": {
        "type": "object",
        "required": ["input"],
        "additionalProperties": false,
        "properties": {
          "input": { "type": "string", "minLength": 1 },
          "attachments": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "ReorderTasksRequest": {
        "type": "object",
        "required": ["taskIds"],
        "additionalProperties": false,
        "properties": {
          "taskIds": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1
          }
        }
      }
    }
  }
}